requirements.txt
pyTelegramBotAPI==4.29.1
psycopg2-binary==2.9.11
python-dotenv==1.2.1

